app/bit-wallet/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
Flask==1.1.2

# Packages
mysql-connector-python==8.0.23

# Automated tests
pytest==6.2.2
pytest-timeout==1.4.2
coverage==5.4
